Fixed errors and temporarily removed the view and logic folders
