+++ include/character_view.h
@@ -1,23 +1,24 @@
-#ifndef FREDVIEW_H
-#define FREDVIEW_H
+#ifndef CHARACTER_VIEW_H
+#define CHARACTER_VIEW_H
 
-#include "character_view.h"
+class CharacterView {
 
-class FredView : public CharacterView {
-    
 private:
     int health, experience, level;
     int inventory[5]; //Temporary -- Needs to be replaced with Item object
 
 public:
 
-    FredView(int health, int experience, int level);
+	CharacterView() {};
+	CharacterView(int health, int experience, int level);
 
     void setExperience(int);
     void setLevel(int);
 
     int getExperience(void);
     int getLevel(void);
+
+
 };
 
-#endif //FREDVIEW_H
+#endif
