+++ include/fred_view.h
@@ -1,25 +1,23 @@
-#ifndef CHARACTERVIEW_H
-#define CHARACTERVIEW_H
+#ifndef FRED_VIEW_H
+#define FRED_VIEW_H
 
-class CharacterView {
+#include "character_view.h"
 
+class FredView : public CharacterView {
+    
 private:
     int health, experience, level;
     int inventory[5]; //Temporary -- Needs to be replaced with Item object
 
-
-
 public:
 
-	CharacterView() {};
-	CharacterView(int health, int experience, int level);
+    FredView(int health, int experience, int level);
 
     void setExperience(int);
     void setLevel(int);
 
     int getExperience(void);
     int getLevel(void);
-
-
 };
-#endif // CHARACTERVIEW_H
+
+#endif //FREDVIEW_H
