+++ src/character.cpp
