+++ src/character_view.cpp
@@ -1,5 +1,5 @@
 #include <SFML/Graphics.hpp>
-#include "view/character_view.h"
+#include "character_view.h"
 
 CharacterView::CharacterView(int health, int experience, int level) {
     this->health = health;
