+++ src/item.cpp
@@ -1,13 +1,12 @@
 #include <SFML/Graphics.hpp>
-#include <logic/item.h>
-#include <logic/actor.h>
+#include "item.h"
+#include "actor.h"
 
 Item::Item(int x, int y, int quantity, bool stackable) {
 	this->x = x;
 	this->y = y;
     this->quantity = quantity;
     this->stackable = stackable;
-
 }
 
 bool Item::isStackable(void) {
